+++ tests/testthat/test-wflow_quickstart.R
@@ -305,3 +305,32 @@ test_that("wflow_quickstart only deletes directory on error if it exists", {
   )
   expect_false(fs::dir_exists(path))
 })
+
+test_that("wflow_quickstart does not accept organization for hosting on GitLab", {
+
+  rmd <- fs::file_temp(ext = ".Rmd")
+  fs::file_create(rmd)
+  on.exit(fs::file_delete(rmd), add = TRUE)
+
+  expect_error(
+    wflow_quickstart(rmd, organization = "gitlab-group", host = "gitlab"),
+    "Do not use the argument"
+  )
+})
+
+test_that("wflow_quickstart fails early if both username and organization are set", {
+
+  rmd <- fs::file_temp(ext = ".Rmd")
+  fs::file_create(rmd)
+  on.exit(fs::file_delete(rmd), add = TRUE)
+  directory <- "test"
+
+  expect_error(
+    wflow_quickstart(rmd, username = "personal-account",
+                     organization = "github-org", directory = directory),
+    "Cannot set both username and organization."
+  )
+
+  # Confirm that directory wasn't created
+  expect_false(fs::dir_exists(directory))
+})